chore: move pkg/sypgp to internal/pkg/sypgp
